Moves refreshing completedReps to jobs dir
