+++ server.js
@@ -5,7 +5,7 @@ const {errorHandler} = require('./middleware/errorMiddleware')
 const connectDB = require('./config/db')
 const port = process.env.PORT || 5000
 const cors = require('cors')
-const refreshCompletedReps = require('./models/scheduler') 
+const refreshCompletedReps = require('./jobs/refreshCompletedReps') 
 
 connectDB();
